Use compiler warnings, and enable them by default
The build system previously installed the gnulib manywarnings module, and
set it up in `configure`, but did not actually use the results to add
warnings flags to the C and C++ compiler invocations in the various
Makefile.am’s.

Fix this!

Also, change the sense of the configure option to be on by default.