feat: Add 'path test' functionality
Both the EPOC16 and EPOC32 remote file servers support path test functionality---`RFSV16_PATHTEST` (`0x22`)  and `RFSV32_PATH_TEST` (`0x2b`) respectively---but this has never been implemented or used internally in plptools. This change introduces a new `rfsv::pathtest` API and provides implementations in `rfsv16` and `rfsv32`.

Psion's implementation of this test is quite counter-intuitive (see the documentation in the header), but it's very helpful when providing a lightweight higher level 'exists' API when combined with `fgetattr`.

While not strictly necessary for plptools in its current implementation, Reconnect relies on this functionality when using plptools as a client library and it makes sense to ensure plptools can provide as complete a protocol implementation as possible. (Ultimately, I hope we can formalize a plptools client API.)

This has been tested on EPOC16 and EPOC32 and behaviour matches that documented in in `rfsv.h`.